# docs/docs/swagger.yaml
basePath: /
definitions:
...
  internal_transport_http.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据载荷
      msg:
        description: 中文提示信息
        type: string
    type: object
...
  tempmail_backend_internal_domain.Webhook:
    properties:
      createdAt:
        type: string
      events:
        description: 订阅的事件类型
        items:
          type: string
        type: array
      id:
        type: string
      isActive:
        description: 是否启用
        type: boolean
      lastError:
        description: 最后一次错误信息
        type: string
      lastSuccess:
        description: 最后一次成功时间
        type: string
      retryCount:
        description: 失败重试次数
        type: integer
      secret:
        description: HMAC 签名密钥
        type: string
      updatedAt:
        type: string
      url:
        description: 回调 URL
        type: string
      userId:
        description: 所属用户ID
        type: string
    type: object
...
  tempmail_backend_internal_service.CreateWebhookInput:
    properties:
      description:
        maxLength: 200
        type: string
      events:
        items:
          type: string
        minItems: 1
        type: array
      url:
        type: string
    required:
    - events
    - url
    type: object
...
  tempmail_backend_internal_service.UpdateWebhookInput:
    properties:
      description:
        maxLength: 200
        type: string
      events:
        items:
          type: string
        minItems: 1
        type: array
      isActive:
        type: boolean
      url:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: TempMail 后端 API 文档
  title: TempMail Backend API
  version: 0.9.0
paths:
...
  /v1/webhooks:
    get:
      consumes:
      - application/json
      description: 列出当前用户的所有 Webhooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/tempmail_backend_internal_domain.Webhook'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 列出 Webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: 创建一个新的 Webhook 配置
      parameters:
      - description: Webhook 信息
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/tempmail_backend_internal_service.CreateWebhookInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/tempmail_backend_internal_domain.Webhook'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 创建 Webhook
      tags:
      - Webhooks
  /v1/webhooks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的 Webhook
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 删除 Webhook
      tags:
      - Webhooks
    get:
      consumes:
      - application/json
      description: 获取指定 Webhook 的详细信息
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/tempmail_backend_internal_domain.Webhook'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 获取 Webhook
      tags:
      - Webhooks
    patch:
      consumes:
      - application/json
      description: 更新 Webhook 配置
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/tempmail_backend_internal_service.UpdateWebhookInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_transport_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/tempmail_backend_internal_domain.Webhook'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 更新 Webhook
      tags:
      - Webhooks
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 兼容层 API Key
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 使用格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
  MailboxToken:
    description: 邮箱访问令牌
    in: header
    name: X-Mailbox-Token
    type: apiKey
swagger: "2.0"